+++ src/UI/WTreeManager.cs
@@ -1,5 +1,3 @@
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using Serilog;
 using WForest.Devices;
 using WForest.Rendering;
@@ -14,21 +12,21 @@ namespace WForest.UI
     /// </summary>
     public class WTreeManager
     {
-        private readonly IWidget _root;
+        public readonly IWidget Root;
         private UserInteractionHandler _userInteractionHandler;
-        private IRenderer _renderer;
-        internal WTreeManager(IWidget widgetRoot, IRenderer renderer)
+
+        internal WTreeManager(IWidget widgetRoot)
         {
             _userInteractionHandler = new UserInteractionHandler(MouseDevice.Instance, new InteractionUpdater());
-            _root = widgetRoot;
-            _renderer = renderer;
-            TreeVisitor.ApplyPropsOnTree(_root);
+            Root = widgetRoot;
+            TreeVisitor.ApplyPropsOnTree(Root);
 
             Log.Information(
                 "Created new WTreeManager with root starting at ({X},{Y}) for ({W},{H}) of space",
-                _root.Space.X, _root.Space.Y, _root.Space.Width, _root.Space.Height
-                );
+                Root.Space.X, Root.Space.Y, Root.Space.Width, Root.Space.Height
+            );
         }
+
 
         /// <summary>
         /// Resize the space in which the WidgetTree is located.
@@ -37,8 +35,8 @@ namespace WForest.UI
         /// <param name="height"></param>
         public void Resize(int width, int height)
         {
-            _root.Space = new RectangleF(_root.Space.X, _root.Space.Y, width, height);
-            TreeVisitor.ApplyPropsOnTree(_root);
+            Root.Space = new RectangleF(Root.Space.X, Root.Space.Y, width, height);
+            TreeVisitor.ApplyPropsOnTree(Root);
         }
 
         /// <summary>
@@ -47,22 +45,20 @@ namespace WForest.UI
         /// </summary>
         public void Update()
         {
-            var transitions = _userInteractionHandler.UpdateAndGenerateTransitions(_root);
+            var transitions = _userInteractionHandler.UpdateAndGenerateTransitions(Root);
             foreach (var t in transitions) t.Execute();
 
-            TreeVisitor.UpdateTree(_root);
+            TreeVisitor.UpdateTree(Root);
         }
 
 
         /// <summary>
         /// Draws the WidgetTree. It visits the entire tree calling the Draw methods of the widgets.
         /// </summary>
-        /// <param name="spriteBatch"></param>
-        public void Draw() => DrawWidgetTree(_renderer);
-
-        private void DrawWidgetTree(IRenderer renderer)
+        /// <param name="renderer"></param>
+        public void Draw(IRenderer renderer)
         {
-            TreeVisitor.ApplyToTreeLevelByLevel(_root, widgets => widgets.ForEach(w =>
+            TreeVisitor.ApplyToTreeLevelByLevel(Root, widgets => widgets.ForEach(w =>
             {
                 w.Draw(renderer);
                 w.PostDrawActions.ForEach(postDraw => postDraw(renderer));